// src/interfaces/MessageObj.ts
export interface MessageObj {
    index:         number;
    guid:          string;
    balance:       string;
    picture:       string;
    name:          string;
    company:       string;
    email:         string;
    phone:         string;
    registered:    string;
    latitude:      number;
    longitude:     number;
    message:       string;
    favoriteFruit: string;
}